separate projects schema
